fix: read fields config from file
