+++ tools/config_field.go
@@ -1,0 +1,6 @@
+package tools
+
+type FieldConfigs []*FieldConfig
+
+type FieldConfig struct {
+}
